Added dev folder with first web page index.html
Why: Well, first of all, we can't have a web site without content. This is the first step in that direction. If we were to run http-server on the dev directory, we would have the very minimum to run a web site.

The name "index.html" is a convention and is one of a number of file names that will be run by default if its folder is opened within a web server

Normally, we will actually run our website from a different folder like app or public. In the future, we will create those as targets for our combined dev code. That way we can develop in a way that is easy to maintain, but later compile that code into something that is more efficient for the server and web-browser to consume.