Added package.json with some simple defaults
Why: package.json is used by npm (node package manager) to automatically pull and install dependency packages when the command "npm install" is used in the same directory as the package file.

Additionally, when we set up grunt to handle more of the heavy lifting needed to build our website from our development files, grunt is often set up to read this file to help it hook up the proper dependencies as well.

This package.json file was created by running "npm init" and answering the questions.