Added README.md documentation
Why: Mainly because npm will get cranky if you don't, even though it doesn't really matter.

Anyway, a README file is basically a file that people will look at first to find out more about what your project is and how to use it.

Additionally, services like github and bitbucket will automatically show your README.md file on the front page of your repository as a sort of landing page.

The .md is short for Markdown, which is a terse syntax for decorating a text file so that it can be shown with formatting in the right viewers. It was designed to look good and readable both in its formatted and raw state.