Add node_modules directory to .gitignore file
Why: now that we have our package.json file and npm is pretty much set up for our current project, we will want to start adding tools that we intend to use to help us develop.

When we add references to these tools in our package.json and run "npm install" on our project directory, npm will put the tools into a folder called "node_modules". There is usually no need to include the actual code for these modules in our source control, so we will have git ignore the directory. The only part we will really need to keep track of are the changes to our package.json file.