Added grunt to package.json devDependencies
Why: grunt is a tool that will help us to automate many of the tasks we would normally need to do in order to make our final output website from our dev files. Grunt can also simplify testing, continuous integration and a number of other tasks that are routine and tedious.

it was added using "npm install grunt --save-dev"

The --save-dev switch caused the package.json file to be updated with the new devDependency.

a devDependency is a dependency that does not need to be in the final output if this package ends up being setup as a downloadable package.