Added Gruntfile.coffee with default task to clean the public folder
Why: Eventually, we will use grunt to build a public folder. Prior to building it, however, we want to make sure it is emptied (this way, if the new build has fewer files, there aren't any files left behind).

The clean task removes the target and all of its contents.

Gruntfile is the default name of the file that grunt looks for when you call in in the project directory. It can be a .js (JavaScript) or .coffee (CoffeeScript) file.

CoffeeScript was used primarily in order to provide some examples of coffeescript usage. Also coffeescript simplifies the task of iterating over all the task names, finding the ones that start with 'grunt-' and loading them into grunt