Added the /public folder to the .gitignore file
Why: Since the "public" directory is a built directory, it makes little sense to add it to source control, as it is basically a copy of what is already there in the dev folder.

It's possible that some build outputs ought to be put into source control, particularly if they are labeled/versioned builds, but that will likely be a different task that outputs to a different folder.