Modified to handle both prod and dev builds
This includes functionality that allows for templating copied files to reflect task information

Why: There are often differences in the way code should be put together depending on whether it is development or production (or even staging or other types of code output). With the processContent option, the copy task is able to integrate these changes into the copied code itself.