Added twitter bootstrap
Set up so that only non-minified files are put into build when dev task is run and only minified files when prod task is run.

Why: In production, we only want the minimum sized css and javascript files, as these must be sent to the client. Whereas, when in development, we want to be able to easily debug these files, therefore, we don't want the minimized files in those cases. Additionally, we could have set things up to have prod point to files in a content delivery network instead of copying those files.

Also notice. In this case, all 3rd party files were put into a directory called libs. A folder was created for the lib name and then a folder for the current version. Finally, all css, img and js files that belong to the lib are kept together within the version folder

Why: There are a number of different philosophies on how this should be handled, however, in my own experience, it has always been easiest when I can keep 3rd party libraries together by version, rather than integrating them into the rest of the application and attempting to locate them when something changes. Of course the gruntfile could be configured to copy these files into a different structure if desired.