Added grunt watch task to package.json
Why: With grunt, this makes it possible to watch one or more files or directories and automatically run tasks whenever those items change. This is particularly useful during development when it is annoying to have to constantly run grunt every time you change a file to see what the output looks like. Of course, once a project gets large, this may make less sense, or it may be necessary to have builds broken into smaller units along with smaller watches.

This was added using "npm install grunt-contrib-watch --save-dev"